Conflicts:
	bdos/bdosmain.c
	bios/kprint.c
	desk/deskwin.c
	include/dos.h
	util/gemdos.c
	util/memmove.S
	util/memset.S
	util/nlsasm.S
	util/setjmp.S
	util/stringasm.S

+++ bdos/bdosmain.c
@@ -396,7 +396,7 @@ static void offree(DMD *d)
 /*
  *  osif -
  */
-long __CDECL osif(int *pw)
+long __CDECL osif(short *pw)
 {
     char **pb, *pb2, *p, ctmp;
     BPB *b;
